sing in donn

+++ app/sign/page.tsx
@@ -4,9 +4,12 @@ import React, { useState } from "react";
 import { RiEye2Fill, RiEyeCloseFill } from "react-icons/ri";
 import { RingLoader } from "react-spinners";
 
-function Login() {
+function Sign() {
+  const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
+
   const [passwordStat, setPasswordStat] = useState("password");
   const [onLogin, setOnLogin] = useState(false);
   const [error, setError] = useState("");
@@ -20,6 +23,14 @@ function Login() {
   }
 
   function hundelSubmi() {
+    if (!name) {
+      setError("Enter name");
+      return;
+    }
+    if (name.length < 5) {
+      setError("name length must upper 5");
+      return;
+    }
     if (!email.match(/^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/)) {
       setError("Enter valid email");
       return;
@@ -28,11 +39,21 @@ function Login() {
       setError("Enter password");
       return;
     }
+    if (password.length < 8) {
+      setError("password length must upper 8");
+      return;
+    }
+    if (password !== confirmPassword) {
+      setError("Confirm password not correct");
+      return;
+    }
     setOnLogin(true);
     axios
-      .post("https://todo-list-servar.onrender.com/auth/login", {
+      .post("https://todo-list-servar.onrender.com/auth/signin", {
+        name,
         email,
         password,
+        passwordConfirm: confirmPassword,
       })
       .then((res) => {
         localStorage.setItem("token", res.data.token);
@@ -41,8 +62,13 @@ function Login() {
         location.replace("/");
       })
       .catch((err) => {
-        if (err.response.data.errors == "wrong password and email") {
+        if (err.response.data.errors === "wrong password and email") {
           setError("Wrong password and email");
+        } else if (
+          err.response.data.message &&
+          err.response.data.message.startsWith("E11000")
+        ) {
+          setError("This email is already in use");
         } else {
           setError("Unknown Error !!!");
         }
@@ -52,7 +78,20 @@ function Login() {
 
   return (
     <form className="flex min-h-screen flex-col  justify-around items-center p-12 sm:p-24  bg-black">
-      <h1 className="text-5xl">Login</h1>
+      <h1 className="text-5xl">Sign in</h1>
+      <div className="flex w-full relative flex-col text-2xl ">
+        <label htmlFor="name">Name : </label>
+        <input
+          required
+          value={name}
+          onChange={(e) => {
+            setName(e.target.value);
+          }}
+          type="text"
+          className="bg-transparent flex-1 border-b-2 focus:outline-none"
+          id="name"
+        />
+      </div>
       <div className="flex w-full relative flex-col text-2xl ">
         <label htmlFor="email">Email : </label>
         <input
@@ -66,7 +105,7 @@ function Login() {
           id="email"
         />
       </div>
-      <div className="flex w-full relative flex-col  text-2xl ">
+      <div className="flex w-full relative flex-col text-2xl ">
         <label htmlFor="password">Password : </label>
         <input
           required
@@ -89,6 +128,19 @@ function Login() {
             onClick={hundelPasswordVea}
           />
         )}
+      </div>
+      <div className="flex w-full relative flex-col text-2xl ">
+        <label htmlFor="confirmPassword">Confirm Password : </label>
+        <input
+          required
+          value={confirmPassword}
+          onChange={(e) => {
+            setConfirmPassword(e.target.value);
+          }}
+          type="text"
+          className="bg-transparent flex-1 border-b-2 focus:outline-none"
+          id="confirmPassword"
+        />
       </div>
       <div className="text-xl text-red-800">{error}</div>
       <div className=" w-full flex justify-center items-center h-16 ">
@@ -113,4 +165,4 @@ function Login() {
 // assemt@ggr.ce
 // Assem1010114101011101000110101110101
 
-export default Login;
+export default Sign;
